chore(core): separated the logic
