+++ examples/mcp-sdk/src/index.ts
@@ -1,4 +1,4 @@
-import { Hono, type Context } from "hono";
+import { Hono } from "hono";
 import {
   muppet,
   describeTool,
@@ -14,9 +14,8 @@ import z from "zod";
 import pino from "pino";
 // import { SSEServerTransport } from "@modelcontextprotocol/sdk/server/sse.js";
 // import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
-import { SSEServerTransport } from "./transport";
+import { SSEHonoTransport, streamSSE } from "muppet/transports";
 import { serve, type HttpBindings } from "@hono/node-server";
-import { SSEStreamingApi } from "hono/streaming";
 
 const app = new Hono();
 const logger = undefined;
@@ -152,7 +151,7 @@ const mcpServer = muppet(app, {
 /**
  * For SSE transport
  */
-let transport: SSEServerTransport | null = null;
+let transport: SSEHonoTransport | null = null;
 
 const server = new Hono<{ Bindings: HttpBindings }>().use(async (c, next) => {
   console.log("Request received", c.req.method, c.req.url);
@@ -162,65 +161,9 @@ const server = new Hono<{ Bindings: HttpBindings }>().use(async (c, next) => {
 
 server.get("/", (c) => c.text("ok!"));
 
-const run = async (
-  stream: SSEStreamingApi,
-  cb: (stream: SSEStreamingApi) => Promise<void>,
-  onError?: (e: Error, stream: SSEStreamingApi) => Promise<void>,
-): Promise<void> => {
-  try {
-    await cb(stream);
-  } catch (e) {
-    if (e instanceof Error && onError) {
-      await onError(e, stream);
-
-      await stream.writeSSE({
-        event: "error",
-        data: e.message,
-      });
-    } else {
-      console.error(e);
-    }
-  }
-};
-
-const contextStash: WeakMap<ReadableStream, Context> = new WeakMap<
-  ReadableStream,
-  Context
->();
-
-const streamSSE = (
-  c: Context,
-  cb: (stream: SSEStreamingApi) => Promise<void>,
-  onError?: (e: Error, stream: SSEStreamingApi) => Promise<void>,
-): Response => {
-  const { readable, writable } = new TransformStream();
-  const stream = new SSEStreamingApi(writable, readable);
-
-  // Until Bun v1.1.27, Bun didn't call cancel() on the ReadableStream for Response objects from Bun.serve()
-  // if (isOldBunVersion()) {
-  //   c.req.raw.signal.addEventListener('abort', () => {
-  //     if (!stream.closed) {
-  //       stream.abort()
-  //     }
-  //   })
-  // }
-
-  // in bun, `c` is destroyed when the request is returned, so hold it until the end of streaming
-  contextStash.set(stream.responseReadable, c);
-
-  c.header("Transfer-Encoding", "chunked");
-  c.header("Content-Type", "text/event-stream");
-  c.header("Cache-Control", "no-cache");
-  c.header("Connection", "keep-alive");
-
-  run(stream, cb, onError);
-
-  return c.newResponse(stream.responseReadable);
-};
-
 server.get("/sse", async (c) => {
   return streamSSE(c, async (stream) => {
-    transport = new SSEServerTransport("/messages", stream);
+    transport = new SSEHonoTransport("/messages", stream);
 
     const mcp = await mcpServer;
     if (!mcp) {
@@ -236,12 +179,12 @@ server.get("/sse", async (c) => {
 });
 
 server.post("/messages", async (c) => {
-  return streamSSE(c, async (stream) => {
-    if (transport) {
-      console.log("Handling post message");
-      await transport.handlePostMessage(c, stream);
-    }
-  });
+  if (!transport) {
+    throw new Error("Transport not initialized");
+  }
+
+  await transport.handlePostMessage(c);
+  return c.text("ok");
 });
 
 server.onError((err, c) => {
